test: Move test files into top-level "test" folder
