+++ rollup.tests.js
@@ -2,6 +2,6 @@ import multiEntry from 'rollup-plugin-multi-entry';
 import babel from 'rollup-plugin-babel';
 
 export default {
-  entry: 'test/lib/*.js',
+  entry: 'test/*.js',
   plugins: [ babel(), multiEntry() ],
 };
